# swagger/ramev2010-delilah_resto-1.0.0-resolved.yaml
...
swagger: "2.0"
info:
  description: "This is the OpenAPI specification for the Delilah Restó system"
  version: "1.0.0"
  title: "Delilah Restó"
  contact:
    email: "2010espinosa.e@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/ramev2010/delilah_resto/1.0.0"
...
schemes:
- "https"
paths:
...
  /users/{username}:
    get:
      tags:
      - "user"
      summary: "Fetches an specific user"
      description: "A determined user can be fetched as well as its complete information.\
        \ Only users with an \"Administrator\" role can fetch users."
      operationId: "fetchUser"
      parameters:
      - name: "username"
        in: "path"
        description: "The desired user's username"
        required: true
        type: "string"
      - name: "authorization"
        in: "header"
        description: "Bearer Type JSON Web Token"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  User:
    type: "object"
    required:
    - "userAddress"
    - "userEmail"
    - "userFirstName"
    - "userLastName"
    - "userPassword"
    - "userPhoneNumber"
    - "userRole"
    - "username"
    properties:
      userFirstName:
        type: "string"
        example: "Ramón"
      userLastName:
        type: "string"
        example: "Espinosa"
      userEmail:
        type: "string"
        format: "email"
        example: "2010espinosa.e@gmail.com"
      userPhoneNumber:
        type: "string"
        example: "3168683691"
      userAddress:
        type: "string"
        example: "Av Cll 128b # 45d-01"
      username:
        type: "string"
        example: "r.espinosa2010"
      userPassword:
        type: "string"
        format: "password"
        example: "myPa$sWord"
      userRole:
        type: "string"
        example: "Administrator"
...
  ErrorResponse:
    type: "object"
    properties:
      errordescription:
        type: "string"
        example: "Unable to create user"
      errorDetail:
        type: "string"
        example: "There is already an user registered with this email"